pyproject.toml
[project]
name = "smart-calendar-agent"
version = "0.1.0"
description = "Find the best experts to consult based on task matching and Google Calendar availability"
requires-python = ">=3.12"
dependencies = [
    "streamlit>=1.40.0",
    "anthropic>=0.40.0",
    "google-api-python-client>=2.150.0",
    "google-auth-oauthlib>=1.2.0",
    "google-auth-httplib2>=0.2.0",
    "pandas>=2.2.0",
    "python-dotenv>=1.0.0",
]

[build-system]
requires = ["setuptools>=75.0"]
build-backend = "setuptools.backends._legacy:_Backend"
